+++ src/ts/errors.ts
@@ -3,5 +3,4 @@ export { PanicErr } from "./errors/panic_err";
 export { UnknownErr } from "./errors/unknown_err";
 export { IoErr } from "./errors/io_err";
 export { InitErr } from "./errors/init_err";
-export { Result } from "./errors/Result";
-export { panic, wrapErr } from "./errors/utils";
+export { Result, panic, wrapErr } from "./errors/utils";
