--- src/ts/errors/Result.ts
@@ -1,3 +1,0 @@
-import { Err } from "./base_err";
-
-export type Result<V, E extends Err> = [V, null] | [null, E];
