--- src/ts/rest/methods.ts
@@ -1,11 +1,0 @@
-export const httpMethods = [
-  "get",
-  "head",
-  "post",
-  "put",
-  "delete",
-  "connect",
-  "options",
-  "trace",
-  "patch",
-] as const;
