+++ src/ts/rest/routing/route.ts
@@ -1,18 +1,37 @@
 import { Application, NextFunction } from "express";
-import { Result } from "../../errors/Result";
+import { Result } from "../../errors";
 import { ApiErr } from "../api_err";
-import { httpMethods } from "../methods";
-import { Request } from "../Request";
 import { RequestContext } from "../request_context";
-import { Response } from "../Response";
+import { Request as ExpressRequest } from "express";
+import { Response as ExpressResponse } from "express";
+import { methods } from "../utils";
+
 export type Handler<Req, Res> = (
   ctx: RequestContext,
   req: Req
 ) => Promise<Result<Res, ApiErr>>;
 
+export interface Request<
+  ResBody = Record<string, unknown>,
+  ReqBody = Record<string, unknown>
+> extends ExpressRequest<
+    Record<string, string>,
+    ResBody,
+    ReqBody,
+    qs.ParsedQs,
+    Record<string, unknown>
+  > {
+  ctx?: RequestContext;
+}
+
+export interface Response<ResBody = Record<string, unknown>>
+  extends ExpressResponse<ResBody, Record<string, unknown>> {
+  body?: ResBody;
+}
+
 export function route<Req, Res>(
   app: Application,
-  method: typeof httpMethods[number],
+  method: typeof methods[number],
   path: string,
   handler: Handler<Req, Res>
 ) {
